% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegmentedHypTestEngine.R
\name{average_power}
\alias{average_power}
\title{average_power.SegmentedHypTestEngine}
\usage{
average_power(
  segmented_hyp_test_engine,
  max_n_segments,
  n_per_segment,
  alpha_total,
  alpha_strong,
  stat_procedure,
  effects_list
)
}
\arguments{
\item{segmented_hyp_test_engine}{A SegmentedHypTestEngine instance for method
dispatch}

\item{max_n_segments}{Maximum number of segments to be run before termination}

\item{n_per_segment}{Number of subjects to be tested in each segment}

\item{alpha_total}{Desired total Type 1 error probability across the entire
experiment}

\item{alpha_strong}{For decision rule: If p-observed for a segment is less
than alpha strong, reject and stop.}

\item{stat_procedure}{Character code for hypothesis testing procedure to
be used in the experiment. Currently must be one of {'1t', '2t', 'r', '1z',
'2z'}}

\item{effects_list}{List of TrueEffect instances}
}
\value{
Average power across all studies based on effects in effect_list
}
\description{
\code{average_power.SegmentedHypTestEngine} computes the average power across
a set of studies as defined by the effects_list argument. Needs to run from
the raw data, not SegHTResults, as SegHTResults do not carry their original
true effect size(s). Called by user-facing methods in SegmentedHypTester.R.
See source code for logic.
}
