% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatProcedureBase.R
\name{critical_value}
\alias{critical_value}
\title{critical_value}
\usage{
critical_value(stat_procedure, alpha_one_tailed, sample_n)
}
\arguments{
\item{stat_procedure}{Instance for method dispatch. Member of StatProcedureBase family}

\item{alpha_one_tailed}{Alpha level for one-tailed test}

\item{sample_n}{Group sizes. Must be a vector for tests with multiple groups}
}
\value{
Critical value of test for alpha and n
}
\description{
Abstract method defined in StatProcedureBase. Should be overridden by all children
}
\details{
\code{critical_value} computes the critical value of a test for given alpha and n
}
