% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatProcedureBase.R
\name{effect_size_from_stat}
\alias{effect_size_from_stat}
\title{effect_size_from_stat}
\usage{
effect_size_from_stat(stat_procedure, stat_value, sample_n)
}
\arguments{
\item{stat_procedure}{Instance for method dispatch. Member of
StatProcedureBase family}

\item{stat_value}{Value of test statistic}

\item{sample_n}{Group sizes. Must be a vector for tests with multiple groups}
}
\value{
Computed effect size
}
\description{
Abstract method defined in StatProcedureBase. Should be overridden by all
children
}
\details{
\code{effect_size_from_stat} computes the effect size for a test given alpha,
n, and an observed value of the test statistic
}
