% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatProcedureBase.R
\name{expected_significant_effect}
\alias{expected_significant_effect}
\title{expected_significant_effect}
\usage{
expected_significant_effect(
  stat_procedure,
  alpha_one_tailed,
  sample_n,
  effect_size
)
}
\arguments{
\item{stat_procedure}{Instance for method dispatch. Member of StatProcedureBase family}

\item{alpha_one_tailed}{Alpha level for one-tailed test}

\item{sample_n}{Group sizes. Must be a vector for tests with multiple groups}

\item{effect_size}{Presumed true effect size}
}
\value{
Expected (mean) effect size
}
\description{
Abstract method defined in StatProcedureBase. Should be overridden by all
children
}
\details{
\code{expected_significant_effect} is used for simulations to demonstrate
bias in published observed effect sizes. Determines the expected (mean) effect
size, if one considers only those cases where the null hypothesis is rejected
(i.e. the observed stat is greater than the critical). For each test, this is
computed by integrating across stat * stat_density for all values above the
critical to obtain an expected (mean) value for that portion of the
distribution, than dividing by the total area under the curve to obtain the
correct proportional value. See child code.
}
