% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatProcedureBase.R
\name{expected_significant_effect_bounded}
\alias{expected_significant_effect_bounded}
\title{expected_significant_effect_bounded}
\usage{
expected_significant_effect_bounded(
  stat_procedure,
  alpha_strong,
  alpha_weak,
  sample_n,
  effect_size
)
}
\arguments{
\item{stat_procedure}{Instance for method dispatch. Member of StatProcedureBase family}

\item{alpha_strong}{Upper bound in sampling distribution}

\item{alpha_weak}{Lower bound in sampling distribution}

\item{sample_n}{Group sizes. Must be a vector for tests with multiple groups}

\item{effect_size}{Presumed true effect size}
}
\value{
Expected (mean) effect size
}
\description{
Abstract method defined in StatProcedureBase. Should be overridden by all
children
}
\details{
\code{expected_significant_effect_bounded} as for expected_significant_effect but for
a specified portion of the sampling distribution
}
