% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatProcedureOneSampleT.R
\name{OneSampleT}
\alias{OneSampleT}
\title{OneSampleT Constructor}
\usage{
OneSampleT(
  min_sample_n = 1,
  display_abbrev = "1t",
  display_name = "1-sample t-test"
)
}
\arguments{
\item{min_sample_n}{Smallest n allowed for this test. Supplied in constructor.}

\item{display_abbrev}{Two-character abbreviation for this test. Developers see one_sample_t_factory.SegmentedHypTester}

\item{display_name}{String value useful for generating tidy outputs for this test}
}
\value{
Validated OneSampleT instance
}
\description{
OneSampleT is a child of StatProcedureBase, representing a one-sample t-test
}
\details{
\code{OneSampleT}  is the public-facing OneSampleT constructor. It creates and returns a class instance
with appropriately initialised values for all fields (min_sample_n, display_abbrev, and
display_name). It requires no parameters.
}
