% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatProcedureBase.R
\name{p_level}
\alias{p_level}
\title{p_level}
\usage{
p_level(stat_procedure, observed_stat, sample_n)
}
\arguments{
\item{stat_procedure}{Instance for method dispatch. Member of StatProcedureBase family}

\item{observed_stat}{Observed test value of interest}

\item{sample_n}{Group sizes. Must be a vector for tests with multiple groups}
}
\value{
Computed p-level
}
\description{
Abstract method defined in StatProcedureBase. Should be overridden by all children
}
\details{
\code{p_level} computes the p_level for an observed statistical values and sample size
}
