% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegmentedHypTestEngine.R
\name{run_scenario.SegmentedHypTestEngine}
\alias{run_scenario.SegmentedHypTestEngine}
\title{run_scenario.SegmentedHypTestEngine}
\usage{
\method{run_scenario}{SegmentedHypTestEngine}(
  segmented_hyp_test_engine,
  segmented_researcher,
  stat_procedure,
  effects_list
)
}
\arguments{
\item{segmented_hyp_test_engine}{A SegmentedHypTestEngine instance for method
dispatch}

\item{segmented_researcher}{Correctly initialised instance of
SegmentedResearcher}

\item{stat_procedure}{Correctly initialised instance of any class descended
from StatProcedureBase, e.g. OneSampleT, PearsonR, etc.}

\item{effects_list}{A list of TrueEffect instances. This must be a list, not
simply a vector. Cast with list() if necessary. Each TrueEffect instance
holds fields effect_size and effect_size_probability. A TrueEffect with
effect_size 0 (i.e. H0 is true) may be included. Probabilities across all effect
sizes must sum to 1 or an exception is thrown.}
}
\value{
A SegmentedHypTestResult instance which holds the average expected outcomes
across all effect sizes, weighted by associated probability.
}
\description{
\code{run_scenario.SegmentedHypTestEngine} numerically determines expected
outcomes for a complete study scenario. It wraps various internal methods of
class SegmentedHypTestEngine (developers can cf. source code).
}
