% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegmentedResearcher.R
\name{SegmentedResearcher}
\alias{SegmentedResearcher}
\title{SegmentedResearcher Cosntructor}
\usage{
SegmentedResearcher(
  max_n_segments = 3,
  n_per_segment = 50,
  alpha_total = 0.05,
  alpha_strong = 0,
  alpha_weak = NA
)
}
\arguments{
\item{max_n_segments}{An integer specifying the maximum number of segments to
be run prior to termination}

\item{n_per_segment}{A \code{numeric} specifying the number of subjects to be
tested in each segment. Note that this value should comprise all subjects
in a segment, summed across any groups.}

\item{alpha_total}{Required Type 1 error probability across the entire study.
Defaults to 0.05}

\item{alpha_strong}{For segmented hypothesis testing decision rule: If
p-observed for any segment is less than alpha_strong, reject H0 and stop
testing.}

\item{alpha_weak}{For decision rule: If p-observed for any segment is
greater than alpha_weak, fail to reject H0 and stop testing. NB: The correct value
of alpha_weak is determined entirely by the values of alpha_total and
alpha_strong and must be derived numerically from those parameters. In most
cases, a value SHOULD NOT be supplied for this argument. When no value is
supplied, the correct value is automatically computed in succeeding
computations.}
}
\value{
Validated SegmentedResearcher instance
}
\description{
A SegmentedResearcher instance holds a collection of values which
define a specific research scenario.
}
\details{
\code{SegmentedResearcher}  is the public-facing SegmentedResearcher constructor. It
provides typical default values for its parameters, but in most cases the user
will supply their own values to define the study scenario they wish to explore.
}
