% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrueEffect.R
\name{TrueEffect}
\alias{TrueEffect}
\title{TrueEffect}
\usage{
TrueEffect(effect_size = 0, effect_size_probability = 1)
}
\arguments{
\item{effect_size}{A \code{numeric} that describes the effect size on a scale from 0 to 1 (e.g. Cohen's d or eta-squared)}

\item{effect_size_probability}{A \code{numeric} between 0 and 1; the base rate of this effect size. Set at 1 if only interested in cases where
the effect is known to exist. Set at less than 1 to allow probability of Ho true.}
}
\value{
Validated TrueEffect instance
}
\description{
Class TrueEffect holds effect size (TrueEffect$effect_size) and probablity (TrueEffect$effect_size_probability),
information needed to describe the true state of the world.
}
\details{
An effect size of 0 implies that the null hypothesis (H0) is true.
An experimental scenario may have one or more true effects, each with a probability of occurrence,
(more generally called a "base rate"). The most typical situations are:

   1) a single effect with probability 1, meaning we consider only those cases where the effect is present

   2) a single effect with probability (base rate) < 1. The researcher expects the effect to be present in this
   proportion of studies across an infinite number of studies in the area. The researcher expects there to be no effect (i.e. H0 to be true) in the remaining 1- base rate
   proportion of studies.

\code{TrueEffect}  is the public-facing TrueEffect constructor. It creates an unvalidated class instance with
 new_TrueEffect, and passes it to validate_TrueEffect. An exception is thrown if either parameter value
 is illegal (outside range {0..1}). It returns the instance if validation does not fail.
}
