% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatProcedureTwoSampleT.R
\name{TwoSampleT}
\alias{TwoSampleT}
\title{TwoSampleT Constructor}
\usage{
TwoSampleT(
  min_sample_n = 4,
  display_abbrev = "2t",
  display_name = "2-sample t-test"
)
}
\arguments{
\item{min_sample_n}{Smallest n allowed for this test. Supplied in constructor.}

\item{display_abbrev}{Two-character abbreviation for this test. Developers see stat_procedure_factory.SegmentedHypTester}

\item{display_name}{String value useful for generating tidy outputs for this test}
}
\value{
Validated TwoSampleT instance
}
\description{
TwoSampleT is a child of StatProcedureBase, representing a two sample t-test
}
\details{
\code{TwoSampleT}  is the public-facing TwoSampleT constructor. It creates and returns a class instance
with appropriately initialised values for all fields (min_sample_n, display_abbrev, and
display_name). It requires no parameters.
}
