% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatProcedureTwoSampleZ.R
\name{TwoSampleZ}
\alias{TwoSampleZ}
\title{TwoSampleZ Constructor}
\usage{
TwoSampleZ(
  min_sample_n = 2,
  display_abbrev = "2z",
  display_name = "2-sample z-test"
)
}
\arguments{
\item{min_sample_n}{Smallest n allowed for this test. Supplied in constructor.}

\item{display_abbrev}{Two-character abbreviation for this test. Developers see stat_procedure_factory.SegmentedHypTester}

\item{display_name}{String value useful for generating tidy outputs for this test}
}
\value{
Validated TwoSampleZ instance
}
\description{
TwoSampleZ is a child of StatProcedureBase, representing a two sample z-test
}
\details{
\code{TwoSampleZ}  is the public-facing TwoSampleZ constructor. It creates and returns a class instance
with appropriately initialised values for all fields (min_sample_n, display_abbrev, and
display_name). It requires no parameters.
}
